pyproject.toml
[project]
name = "drug-rescue"
version = "0.1.0"
description = "AI-powered drug repurposing agent — TreeHacks 2026"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pydantic>=2.8.0",
]

[project.optional-dependencies]
agent = [
    "claude-agent-sdk>=0.1.30",
]
api = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.1",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
drug-rescue = "drug_rescue.agent:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
